requirements/base.txt
django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
django-filter==22.1